Fix a precision issue in test expression

+++ iarray/tests/test_expression.py
@@ -10,7 +10,7 @@ import numpy as np
      ("iterblosc", [1000], [100], np.float64, "(cos(x) - 1.35) * (sin(x) - 4.45) * tan(x - 8.5)"),
      ("iterchunk", [1000], [123], np.float32, "(abs(-x) - 1.35) * ceil(x) * floor(x - 8.5)"),
      ("iterblock", [100, 100], [23, 32], np.float64, "sinh(x) + (cosh(x) - 1.35) - tanh(x + .2)"),
-     ("iterchunk", [100, 100, 55], [10, 5, 10], np.float32, "asin(x) + (acos(x) - 1.35) - atan(x + .2)"),
+     ("iterchunk", [100, 100, 55], [10, 5, 10], np.float64, "asin(x) + (acos(x) - 1.35) - atan(x + .2)"),
      ("iterblock", [1000], None, np.float64, "exp(x) + (log(x) - 1.35) - log10(x + .2)"),
      ("iterchunk", [1000], None, np.float32, "sqrt(x) + atan2(x, x) + pow(x, x)"),
      ("iterblock", [100, 100], None, np.float64, "(x - cos(1)) * 2"),
